% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bolt_lmm.R
\name{bolt_lmm}
\alias{bolt_lmm}
\title{Run BOLT-LMM \code{--lmm} and returns table of association statistics.}
\usage{
bolt_lmm(name, name_out = name, bolt_bin = "bolt", threads = 0, verbose = TRUE)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions (the argument to \code{--bfile}).}

\item{name_out}{The base name of the output statistics file (the argument to \code{--statsFile}; default same as input \code{name}), which gets extension ".bolt.lmm" added automatically.}

\item{bolt_bin}{The path to the binary executable.
Default assumes \code{bolt} is in the PATH.}

\item{threads}{The number of threads to use (the argument to \code{--numThreads}).
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed, followed by the the path of the output file being read (after autocompleting the extensions).}
}
\value{
The table of genetic association statistics, as a \code{tibble}, read with \code{\link[=read_bolt_lmm]{read_bolt_lmm()}} (see that for more info).
}
\description{
A wrapper for running BOLT-LMM's genetic association test followed by reading of the results by \code{\link[=read_bolt_lmm]{read_bolt_lmm()}}.
In addition to the arguments listed below, the executable is run with \code{--phenoUseFam} (so the phenotype tested can only be the one present in the input FAM file), \code{--LDscoresUseChip}, \code{--lmmForceNonInf} (so columns are never missing), and \verb{--maxMissingPerSnp 1 --maxMissingPerIndiv 1} (so full data is used and loci are never missing from table).
}
\examples{
\dontrun{
data <- bolt_lmm( name, name_out )
}

}
\seealso{
\code{\link[=read_bolt_lmm]{read_bolt_lmm()}} for parsing the output of BOLT-LMM.

\code{\link[=delete_files_bolt]{delete_files_bolt()}} for deleting the output of BOLT-LMM.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing BOLT-LMM and error handling.
}
