% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_assoc.R
\name{delete_files_assoc}
\alias{delete_files_assoc}
\title{Delete GEMMA output}
\usage{
delete_files_assoc(file)
}
\arguments{
\item{file}{The file path (excluding extension \code{assoc.txt}).}
}
\value{
Nothing
}
\description{
This function deletes the GEMMA output file (\code{assoc.txt} extension) given the base file path, warning if the file did not exist or if it was not successfully deleted.
}
\examples{
\dontrun{
delete_files_assoc("data")
}

}
\seealso{
\code{\link[=gemma_lmm]{gemma_lmm()}}
}
