% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_bolt.R
\name{delete_files_bolt}
\alias{delete_files_bolt}
\title{Delete BOLT-LMM output}
\usage{
delete_files_bolt(file)
}
\arguments{
\item{file}{The file path (excluding extension \code{bolt.txt}).}
}
\value{
Nothing
}
\description{
This function deletes the BOLT-LMM output file (\code{bolt.txt} extension) given the base file path, warning if the file did not exist or if it was not successfully deleted.
}
\examples{
\dontrun{
delete_files_bolt("data")
}

}
\seealso{
\code{\link[=bolt_lmm]{bolt_lmm()}}
}
