% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_gcta_hsq.R
\name{delete_files_gcta_hsq}
\alias{delete_files_gcta_hsq}
\title{Delete GCTA REML output}
\usage{
delete_files_gcta_hsq(file)
}
\arguments{
\item{file}{The file path (excluding extension \code{hsq}).}
}
\value{
Nothing
}
\description{
This function deletes the GCTA REML output file (\code{hsq} extension) given the base file path, warning if the file did not exist or if it was not successfully deleted.
}
\examples{
\dontrun{
delete_files_gcta_hsq("data")
}

}
\seealso{
\code{\link[=gcta_reml]{gcta_reml()}}
}
