% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_gcta_mlma.R
\name{delete_files_gcta_mlma}
\alias{delete_files_gcta_mlma}
\title{Delete GCTA MLMA output}
\usage{
delete_files_gcta_mlma(file)
}
\arguments{
\item{file}{The file path (excluding extension \code{mlma}).}
}
\value{
Nothing
}
\description{
This function deletes the GCTA MLMA output file (\code{mlma} extension) given the base file path, warning if the file did not exist or if it was not successfully deleted.
}
\examples{
\dontrun{
delete_files_gcta_mlma("data")
}

}
\seealso{
\code{\link[=gcta_mlma]{gcta_mlma()}}
}
