% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_logtxt.R
\name{delete_files_logtxt}
\alias{delete_files_logtxt}
\title{Delete GEMMA LOG file}
\usage{
delete_files_logtxt(file)
}
\arguments{
\item{file}{The file path (excluding extension \code{log.txt}).}
}
\value{
Nothing
}
\description{
This function deletes a GEMMA log file (\code{log.txt} extension) given the base file path, warning if the file did not exist or if it was not successfully deleted.
}
\examples{
\dontrun{
delete_files_logtxt("data")
}

}
