% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_pca.R
\name{delete_files_pca}
\alias{delete_files_pca}
\title{Delete PCA outputs}
\usage{
delete_files_pca(file)
}
\arguments{
\item{file}{The shared file path (excluding extensions: \code{eigenvec}, \code{eigenval}).}
}
\value{
Nothing
}
\description{
This function deletes each of the standard PCA output files (\code{eigenvec}, \code{eigenval} extensions) given the shared base file path, warning if any of the files did not exist or if any were not successfully deleted.
}
\examples{
\dontrun{
delete_files_pca("data")
}

}
\seealso{
\code{\link[=plink_pca]{plink_pca()}}

\code{\link[=delete_files_log]{delete_files_log()}}
}
