% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_plink_glm.R
\name{delete_files_plink_glm}
\alias{delete_files_plink_glm}
\title{Delete \code{plink2 --glm} output}
\usage{
delete_files_plink_glm(file)
}
\arguments{
\item{file}{The file path (excluding extension \code{PHENO1.glm.linear}).}
}
\value{
Nothing
}
\description{
This function deletes the \code{plink2 --glm} output file (\code{PHENO1.glm.linear} extension) given the base file path, warning if the file did not exist or if it was not successfully deleted.
}
\examples{
\dontrun{
delete_files_plink_glm("data")
}

}
\seealso{
\code{\link[=plink_glm]{plink_glm()}}
}
