% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_plink_hardy.R
\name{delete_files_plink_hardy}
\alias{delete_files_plink_hardy}
\title{Delete \code{plink2 --hardy} output}
\usage{
delete_files_plink_hardy(file)
}
\arguments{
\item{file}{The file path (excluding extension \code{hardy}).}
}
\value{
Nothing
}
\description{
This function deletes the \code{plink2 --hardy} output file (\code{hardy} extension) given the base file path, warning if the file did not exist or if it was not successfully deleted.
}
\examples{
\dontrun{
delete_files_plink_hardy("data")
}

}
\seealso{
\code{\link[=plink_hardy]{plink_hardy()}}
}
