% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delete_files_tped.R
\name{delete_files_tped}
\alias{delete_files_tped}
\title{Delete TPED/TFAM files}
\usage{
delete_files_tped(file)
}
\arguments{
\item{file}{The shared file path (excluding extensions: \code{tped}, \code{tfam}).}
}
\value{
Nothing
}
\description{
This function deletes each of the transposed plink output files (\code{tped}, \code{tfam} extensions) given the shared base file path, warning if any of the files did not exist or if any were not successfully deleted.
}
\examples{
\dontrun{
delete_files_tped("data")
}

}
\seealso{
\code{\link[=plink1_bed_to_tped]{plink1_bed_to_tped()}}
}
