% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmax_kin.R
\name{emmax_kin}
\alias{emmax_kin}
\title{Estimate kinship with EMMAX}
\usage{
emmax_kin(name, emmax_kin_bin = "emmax-kin", verbose = TRUE)
}
\arguments{
\item{name}{The shared name of the input plink TPED/TFAM files without extensions.}

\item{emmax_kin_bin}{The path to the binary executable.
Default assumes \code{emmax-kin} is in the PATH.}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
Nothing (the kinship matrix is not read in; see example below for reading with a different function).
}
\description{
Uses \code{emmax-kin} with default option to estimate the kinship matrix.
The output file will be named \code{name}.BN.kinf using the input name below (emmax-kin does not allow the input and output names to be different).
}
\examples{
\dontrun{
# create kinship matrix estimate (on file)
emmax_kin( name )

# read file with genio
kinship <- genio::read_matrix( name, ext = 'BN.kinf' )
}

}
\seealso{
\code{\link[genio:read_matrix]{genio::read_matrix()}} for reading plain-text matrices.

\code{\link[=plink1_bed_to_tped]{plink1_bed_to_tped()}} for creating TPED/TFAM files from BED/BIM/FAM files.

\code{\link[=require_files_tped]{require_files_tped()}} for requiring that the TPED/TFAM files exist (used internally by \code{emmax_kin}).

\code{\link[=delete_files_tped]{delete_files_tped()}} for deleting TPED/TFAM files.

\code{\link[=emmax_lmm]{emmax_lmm()}} for running GWAS with EMMAX.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing EMMAX and error handling.
}
