% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmax_lmm.R
\name{emmax_lmm}
\alias{emmax_lmm}
\title{Run EMMAX and returns table of association statistics.}
\usage{
emmax_lmm(
  name,
  file_kinship,
  name_phen = name,
  name_out = name,
  maf = NULL,
  emmax_bin = "emmax",
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink TPED/TFAM files without extensions.}

\item{file_kinship}{The path to the kinship matrix file.}

\item{name_phen}{The base name of the phenotype file without .phen extension (default same as input \code{name}).}

\item{name_out}{The base name of the output statistics file (default same as input \code{name}), which gets extension ".ps" added automatically.}

\item{maf}{Optional vector of allele frequencies (misnamed, these cannot be \emph{minor} allele frequencies or nothing gets corrected), used to correct signs of regression coefficients (EMMAX reports them in terms of the major allele, not the original allele).  If missing, no corrections are applied.}

\item{emmax_bin}{The path to the binary executable.
Default assumes \code{emmax} is in the PATH.}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
The table of genetic association statistics, as a \code{tibble}, read with \code{\link[=read_emmax_ps]{read_emmax_ps()}} (see that for more info).
}
\description{
A wrapper for running EMMAX genetic association test followed by reading of the results by \code{\link[=read_emmax_ps]{read_emmax_ps()}}.
}
\examples{
\dontrun{
data <- emmax_lmm( name, file_kinship, name_out )
}

}
\seealso{
\code{\link[=plink1_bed_to_tped]{plink1_bed_to_tped()}} for creating TPED/TFAM files from BED/BIM/FAM files.

\code{\link[=require_files_tped]{require_files_tped()}} for requiring that the TPED/TFAM files exist (used internally by \code{emmax_lmm}).

\code{\link[=delete_files_tped]{delete_files_tped()}} for deleting TPED/TFAM files.

\code{\link[=emmax_kin]{emmax_kin()}} for estimating the kinship matrix with EMMAX.

\code{\link[=read_emmax_ps]{read_emmax_ps()}} for parsing the output of EMMAX.

\code{\link[=delete_files_emmax_lmm]{delete_files_emmax_lmm()}} and \code{\link[=delete_files_log]{delete_files_log()}} for deleting the EMMAX output files.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing EMMAX and error handling.
}
