% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcta_grm.R
\name{gcta_grm}
\alias{gcta_grm}
\title{Estimate kinship (GRM) with GCTA}
\usage{
gcta_grm(
  name,
  name_out = name,
  gcta_bin = "gcta64",
  threads = 0,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions.}

\item{name_out}{The base name of the output binary GRM files (default same as input \code{name}), which get extensions (\code{grm.bin}, \code{grm.N.bin}, \code{grm.id}) added automatically.}

\item{gcta_bin}{The path to the binary executable.
Default assumes \code{gcta64} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
Nothing (the kinship matrix is not read in; see example below for reading with a different function).
}
\description{
Uses \code{gcta64} with \code{-grm} and default option to estimate the kinship matrix.
The output file will be named \code{name_out} using the output name below with extensions \code{grm.bin}, \code{grm.N.bin}, and \code{grm.id}.
}
\examples{
\dontrun{
# create kinship matrix estimate (on file)
gcta_grm( name, name_out )

# read file with genio
data <- genio::read_grm( name_out )
kinship <- data$kinship
}

}
\seealso{
\code{\link[genio:read_grm]{genio::read_grm()}} for reading binary GRM matrices.

\code{\link[=gcta_mlma]{gcta_mlma()}} for running GWAS with GCTA.

\code{\link[=gcta_pca]{gcta_pca()}} for estimating PCA from GRM with GCTA.

\code{\link[genio:delete_files_grm]{genio::delete_files_grm()}} for deleting GRM files.

\code{\link[=delete_files_log]{delete_files_log()}} for deleting the GCTA log file.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing GCTA and error handling.
}
