% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcta_mlma.R
\name{gcta_mlma}
\alias{gcta_mlma}
\title{Run GCTA MLMA and returns table of association statistics.}
\usage{
gcta_mlma(
  name,
  name_grm = name,
  name_phen = name,
  name_out = name,
  file_covar = NULL,
  file_covar_cat = NULL,
  gcta_bin = "gcta64",
  threads = 0,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions.}

\item{name_grm}{The shared name of the input binary GRM files without extensions (default same as input \code{name}).}

\item{name_phen}{The base name of the phenotype file without PHEN extension (default same as input \code{name}).}

\item{name_out}{The base name of the output statistics file (default same as input \code{name}), which gets extension ".mlma" added automatically.}

\item{file_covar}{Optional file path of fixed continuous covariates.}

\item{file_covar_cat}{Optional file path of fixed categorical covariates.}

\item{gcta_bin}{The path to the binary executable.
Default assumes \code{gcta64} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed, followed by the the path of the output file being read (after autocompleting the extensions).}
}
\value{
The table of genetic association statistics, as a \code{tibble}, read with \code{\link[=read_gcta_mlma]{read_gcta_mlma()}} (see that for more info).
Certain errors in the binary (see above) cause \code{NULL} to be returned.
}
\description{
A wrapper for running GCTA genetic association test followed by reading of the results by \code{\link[=read_gcta_mlma]{read_gcta_mlma()}}.
If GCTA returns with a non-zero status, the wrapper inspects STDOUT/STDERR and the log file for messages about non-invertible matrices or convergence failures (cases encountered often when an excessive number of covariates is provided), and returns \code{NULL} for those cases without errors or warnings in R (all other failures make the wrapper stop with an error).
}
\examples{
\dontrun{
data <- gcta_mlma( name, name_out )
}

}
\seealso{
\code{\link[=gcta_grm]{gcta_grm()}} for estimating the kinship (GRM) matrix with GCTA.

\code{\link[=read_gcta_mlma]{read_gcta_mlma()}} for parsing the output of GCTA MLMA.

\code{\link[=delete_files_gcta_mlma]{delete_files_gcta_mlma()}} and \code{\link[=delete_files_log]{delete_files_log()}} for deleting the GCTA output files.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing GCTA and error handling.

Input formats for continuous (\code{--qcovar}) and categorical (\code{--covar}) covariate files:
\url{https://yanglab.westlake.edu.cn/software/gcta/#GREMLanalysis}
}
