% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcta_pca.R
\name{gcta_pca}
\alias{gcta_pca}
\title{Estimate principal components with GCTA}
\usage{
gcta_pca(
  name,
  name_out = name,
  n_pcs = 10,
  gcta_bin = "gcta64",
  threads = 0,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input binary GRM files without extensions.}

\item{name_out}{The base name of the output eigenvec/eigenval files (default same as input \code{name}), which gets extensions added automatically.}

\item{n_pcs}{The number of eigenvectors/eigenvalues (a.k.a. principal components) to calculate.}

\item{gcta_bin}{The path to the binary executable.
Default assumes \code{gcta64} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
Nothing (the eigenvec/eigenval files are not read in; see example below for reading with a different function).
}
\description{
A wrapper for running GCTA's PCA.
}
\examples{
\dontrun{
# create PCs (on file)
gcta_pca( name, name_out )

# read file with genio
data <- genio::read_eigenvec( name_out )
}

}
\seealso{
\code{\link[genio:read_eigenvec]{genio::read_eigenvec()}} for reading eigenvec files.

\code{\link[=delete_files_pca]{delete_files_pca()}} for deleting the eigenvec/eigenval outputs.

\code{\link[=delete_files_log]{delete_files_log()}} for deleting GCTA log files.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing GCTA and error handling.

\url{http://cnsgenomics.com/software/gcta/#PCA}
}
