% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcta_reml.R
\name{gcta_reml}
\alias{gcta_reml}
\title{Run GCTA REML and return the estimated heritability.}
\usage{
gcta_reml(
  name,
  name_grm = name,
  name_phen = name,
  name_out = name,
  gcta_bin = "gcta64",
  threads = 0,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared default name of the input and output files without extensions.}

\item{name_grm}{The shared name of the input binary GRM files without extensions (default same as input \code{name}).}

\item{name_phen}{The base name of the phenotype file without PHEN extension (default same as input \code{name}).}

\item{name_out}{The base name of the output variance component file (default same as input \code{name}), which gets extension ".hsq" added automatically.}

\item{gcta_bin}{The path to the binary executable.
Default assumes \code{gcta64} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed, followed by the the path of the output file being read (after autocompleting the extensions).}
}
\value{
A list containing the estimated heritability, its standard error, and the full table of estimated variance components, as a \code{tibble}, read with \code{\link[=read_gcta_hsq]{read_gcta_hsq()}} (see that for more info).
}
\description{
A wrapper for running GCTA variance component estimation followed by reading of the results by \code{\link[=read_gcta_hsq]{read_gcta_hsq()}}.
}
\examples{
\dontrun{
obj <- gcta_reml( name, name_out )
obj$herit
}

}
\seealso{
\code{\link[=gcta_grm]{gcta_grm()}} for estimating the kinship (GRM) matrix with GCTA.

\code{\link[=read_gcta_hsq]{read_gcta_hsq()}} for parsing the output of GCTA MLMA.

\code{\link[=delete_files_gcta_hsq]{delete_files_gcta_hsq()}} and \code{\link[=delete_files_log]{delete_files_log()}} for deleting the GCTA output files.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing GCTA and error handling.
}
