% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma_kin.R
\name{gemma_kin}
\alias{gemma_kin}
\title{Estimate kinship with GEMMA}
\usage{
gemma_kin(
  name,
  name_out = name,
  gemma_bin = "gemma",
  threads = 0,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions.}

\item{name_out}{The base name of the output kinship file (default same as input \code{name}), which gets extension ".cXX.txt" added automatically.}

\item{gemma_bin}{The path to the binary executable.
Default assumes \code{gemma} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
Nothing (the kinship matrix is not read in; see example below for reading with a different function).
}
\description{
Uses \code{gemma} with \code{-gk} and default option to estimate the kinship matrix.
The output file will be named \code{name_out}.cXX.txt using the output name below.
}
\examples{
\dontrun{
# create kinship matrix estimate (on file)
gemma_kin( name )

# read file with genio
kinship <- genio::read_matrix( name, ext = 'cXX.txt' )
}

}
\seealso{
\code{\link[genio:read_matrix]{genio::read_matrix()}} for reading plain-text matrices.

\code{\link[=gemma_lmm]{gemma_lmm()}} for running GWAS with GEMMA.

\code{\link[=delete_files_logtxt]{delete_files_logtxt()}} for deleting the GEMMA log file.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing GEMMA and error handling.
}
