% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gemma_lmm.R
\name{gemma_lmm}
\alias{gemma_lmm}
\title{Run GEMMA and returns table of association statistics.}
\usage{
gemma_lmm(
  name,
  file_kinship,
  name_out = name,
  gemma_bin = "gemma",
  threads = 0,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions.
Phenotype from FAM table is used.}

\item{file_kinship}{The path to the kinship matrix file.}

\item{name_out}{The base name of the output statistics file (default same as input \code{name}), which gets extension ".assoc.txt" added automatically.}

\item{gemma_bin}{The path to the binary executable.
Default assumes \code{gemma} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
The table of genetic association statistics, as a \code{tibble}, read with \code{\link[=read_gemma_assoc]{read_gemma_assoc()}} (see that for more info).
}
\description{
A wrapper for running GEMMA genetic association test followed by reading of the results by \code{\link[=read_gemma_assoc]{read_gemma_assoc()}}.
}
\examples{
\dontrun{
data <- gemma_lmm( name, file_kinship, name_out )
}

}
\seealso{
\code{\link[=gemma_kin]{gemma_kin()}} for estimating the kinship matrix with GEMMA.

\code{\link[=read_gemma_assoc]{read_gemma_assoc()}} for parsing the output of GEMMA.

\code{\link[=delete_files_assoc]{delete_files_assoc()}} and \code{\link[=delete_files_logtxt]{delete_files_logtxt()}} for deleting the GEMMA output files.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing GEMMA and error handling.
}
