% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genbin-package.R
\docType{package}
\name{genbin-package}
\alias{genbin-package}
\alias{genbin}
\title{genbin: R wrappers for binaries in genetics}
\description{
Conveniently run genetics analysis involving external binary executables from R.
Pipelines support simple procedures for creating, reading, and cleaning up files, in some cases complemented by the \code{genio} package.
}
\examples{
\dontrun{
library(genbin)
library(genio)

# examples below assume these "plink" files exist:
# name.bed, name.bim, name.fam, name.phen
name <- 'name'
# number of PCs for some examples:
n_pcs <- 10

### GCTA pipeline

# The GCTA examples assume that `gcta64` is a binary in the system's PATH.

# create GRM from plink files
gcta_grm( name )
# optional: read kinship matrix into R
data <- genio::read_grm( name )
kinship <- data$kinship

# perform mixed linear model association, returning table
data <- gcta_mlma( name )
# association p-values
data$p

# optional: calculate PCs (creates eigenvec/eigenval files)
gcta_pca( name, n_pcs = n_pcs )
# read PCs into R
data <- genio::read_eigenvec( name )
# delete eigenvec/eigenval files
delete_files_pca( name )

# cleanup
# delete GRM files
genio::delete_files_grm( name )
# delete association table
delete_files_gcta_mlma( name )
# delete log
delete_files_log( name )

### PLINK PCA pipeline

# The plink examples assume that `plink2` is a binary in the system's PATH.

# calculate PCs (creates eigenvec/eigenval files)
plink_pca( name, n_pcs = n_pcs )
# optional: read PCs into R
data <- genio::read_eigenvec( name )

# perform PCA association, returning table
data <- plink_glm( name, file_covar = paste0( name, '.eigenvec' ) )
# association p-values
data$p

# cleanup
# delete eigenvec/eigenval files
delete_files_pca( name )
# delete association table
delete_files_plink_glm( name )
# delete log
delete_files_log( name )

}

}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/OchoaLab/genbin}
  \item Report bugs at \url{https://github.com/OchoaLab/genbin/issues}
}

}
\author{
\strong{Maintainer}: Alejandro Ochoa \email{alejandro.ochoa@duke.edu} (\href{https://orcid.org/0000-0003-4928-3403}{ORCID})

}
\keyword{internal}
