% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink1_bed_to_tped.R
\name{plink1_bed_to_tped}
\alias{plink1_bed_to_tped}
\title{Reformat BED/BIM/FAM to TPED/TFAM}
\usage{
plink1_bed_to_tped(
  name,
  name_out = name,
  plink1_bin = "plink1",
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions.}

\item{name_out}{The base name of the output TPED/TFAM files without extensions (default same as input \code{name}).}

\item{plink1_bin}{The path to the binary executable.
Default assumes \code{plink1} is in the PATH.}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
Nothing.
}
\description{
Uses \code{plink1} to perform reformatting (note \code{plink2} does not have this functionality, as tped/tfam is a very old and obsolete format).
}
\examples{
\dontrun{
plink1_bed_to_tped( name )
}

}
\seealso{
\code{\link[=require_files_tped]{require_files_tped()}} for requiring that the TPED/TFAM files exist.

\code{\link[=delete_files_tped]{delete_files_tped()}} for deleting TPED/TFAM files.

\code{\link[=delete_files_log]{delete_files_log()}} for deleting the plink log file.

\code{\link[=emmax_kin]{emmax_kin()}}, \code{\link[=emmax_lmm]{emmax_lmm()}} for functions that require TPED/TFAM files.
}
