% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_glm.R
\name{plink_glm}
\alias{plink_glm}
\title{Run \code{plink2 --glm} and returns table of association statistics.}
\usage{
plink_glm(
  name,
  name_phen = name,
  name_out = name,
  file_covar = NULL,
  plink_bin = "plink2",
  threads = 0,
  ver_older = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions.}

\item{name_phen}{The base name of the phenotype file without PHEN extension (default same as input \code{name}).
Only the first phenotype (column 3) is used.}

\item{name_out}{The base name of the output statistics file (default same as input \code{name}), which gets extension ".PHENO1.glm.linear" added automatically.}

\item{file_covar}{Optional file path of fixed covariates for GLM.
Pass the EIGENVEC file (with extension) to control for population structure using principal components.}

\item{plink_bin}{The path to the binary executable.
Default assumes \code{plink2} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{ver_older}{Handles the case of missing covariates appropriately for older plink2 versions.
In the newer versions we must include \code{allow-no-covars} option or plink2 fails; however in older versions we must not include \code{allow-no-covars} or plink2 fails!}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed, followed by the the path of the output file being read (after autocompleting the extensions).}
}
\value{
The table of genetic association statistics, as a \code{tibble}, read with \code{\link[=read_plink_glm]{read_plink_glm()}} (see that for more info).
}
\description{
A wrapper for running plink2's genetic association test followed by reading of the results by \code{\link[=read_plink_glm]{read_plink_glm()}}.
The \code{--glm} option is run with the modifiers \code{omit-ref} and \code{hide-covar}, which simplifies outputs.
}
\examples{
\dontrun{
data <- plink_glm( name, name_out = name_out )
}

}
\seealso{
\code{\link[=read_plink_glm]{read_plink_glm()}} for parsing the output of \code{plink2 --glm}.

\code{\link[=delete_files_plink_glm]{delete_files_plink_glm()}}, \code{\link[=delete_files_log]{delete_files_log()}} for deleting the output of \code{plink2 --glm}.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing \code{plink2} and error handling.
}
