% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_hardy.R
\name{plink_hardy}
\alias{plink_hardy}
\title{Calculate Hard-Weinberg equilibrium (HWE) p-values and other statistics with plink2}
\usage{
plink_hardy(
  name,
  name_out = name,
  midp = FALSE,
  plink_bin = "plink2",
  threads = 0,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions.}

\item{name_out}{The base name of the output hardy file (default same as input \code{name}), which gets extensions added automatically.}

\item{midp}{If \code{TRUE}, uses mid p-value method.  Default doesn't use it.}

\item{plink_bin}{The path to the binary executable.
Default assumes \code{plink2} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
The table of HWE statistics, as a \code{tibble}, read with \code{\link[=read_plink_hardy]{read_plink_hardy()}} (see that for more info).
}
\description{
A wrapper for plink2's \code{hardy} option.
In addition to the arguments listed below, the executable is run with \code{--silent} and \code{--nonfounders} (to use all individuals whether they are labeled as founders or not).
}
\examples{
\dontrun{
# calculate desired statistics
data <- plink_hardy( name )
}

}
\seealso{
\code{\link[=read_plink_hardy]{read_plink_hardy()}} for reading hardy files.

\code{\link[=delete_files_plink_hardy]{delete_files_plink_hardy()}} for deleting the hardy outputs.

\code{\link[=delete_files_log]{delete_files_log()}} for deleting plink log files.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing plink2 and error handling.
}
