% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plink_pca.R
\name{plink_pca}
\alias{plink_pca}
\title{Estimate principal components with plink2}
\usage{
plink_pca(
  name,
  name_out = name,
  n_pcs = 10,
  maf = 0.1,
  approx = FALSE,
  plink_bin = "plink2",
  threads = 0,
  verbose = TRUE
)
}
\arguments{
\item{name}{The shared name of the input plink BED/BIM/FAM files without extensions.}

\item{name_out}{The base name of the output eigenvec/eigenval files (default same as input \code{name}), which gets extensions added automatically.}

\item{n_pcs}{The number of eigenvectors/eigenvalues (a.k.a. principal components) to calculate.}

\item{maf}{The minor allele frequency threshold to apply to data prior to eigendecomposition.}

\item{approx}{Apply approximate PCA algorithm.}

\item{plink_bin}{The path to the binary executable.
Default assumes \code{plink2} is in the PATH.}

\item{threads}{The number of threads to use.
The values 0 (default), NA, or NULL use all threads available (the output of \code{\link[parallel:detectCores]{parallel::detectCores()}}).}

\item{verbose}{If \code{TRUE} (default), prints the command line before it is executed.}
}
\value{
Nothing (the eigenvec/eigenval files are not read in; see example below for reading with a different function).
}
\description{
A wrapper for running plink2's PCA.
In addition to the arguments listed below, the executable is run with \code{--silent}, \code{--nonfounders} (to use all individuals whether they are labeled as founders or not), and \code{--bad-freqs} (to apply even when sample sizes are very small).
}
\examples{
\dontrun{
# create PCs (on file)
plink_pca( name, name_out )

# read file with genio
data <- genio::read_eigenvec( name_out )
}

}
\seealso{
\code{\link[genio:read_eigenvec]{genio::read_eigenvec()}} for reading eigenvec files.

\code{\link[=delete_files_pca]{delete_files_pca()}} for deleting the eigenvec/eigenval outputs.

\code{\link[=delete_files_log]{delete_files_log()}} for deleting plink log files.

\code{\link[=system3]{system3()}}, used (with \code{ret = FALSE}) for executing plink2 and error handling.
}
