% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bolt_lmm.R
\name{read_bolt_lmm}
\alias{read_bolt_lmm}
\title{Read BOLT-LMM statistics table}
\usage{
read_bolt_lmm(file, ext = "bolt.txt", verbose = TRUE)
}
\arguments{
\item{file}{The file path to read, possibly missing the extension.}

\item{ext}{The expected file extension.
Set to \code{NA} to prevent adding any extension to \code{file}.
BOLT-LMM does not create an output with a standard extension, so I chose the default extension to match what is generated by \code{\link[=bolt_lmm]{bolt_lmm()}}.}

\item{verbose}{If \code{TRUE} (default) function reports the path of the file being loaded (after autocompleting the extensions).}
}
\value{
The table as a \code{tibble}.
Original names (as they appear in the header line of the file) are modified by lowercasing all, followed by these specific mappings to ensure that columns shared by BIM table have the same names as those returned by \code{\link[genio:read_bim]{genio::read_bim()}}, the association p-value is "p", and the regression coefficient is "beta" ("orig" -> "new"):
\itemize{
\item "snp" -> "id"
\item "bp" -> "pos"
\item "genpos" -> "posg"
\item "allele1" -> "ref"
\item "allele0" -> "alt"
\item "p_bolt_lmm_inf" -> "p_inf"
\item "p_bolt_lmm" -> "p"
}
}
\description{
Parses the output generated by BOLT-LMM with option \code{--lmm} (genetic assocation test) and specified with the option \verb{--statsFile <file>}.
Returned table has column names standardized for convenience.
Validated with BOLT-LMM version 2.3.5.
}
\examples{
\dontrun{
data <- read_bolt_lmm( file )
}

}
\seealso{
\code{\link[=bolt_lmm]{bolt_lmm()}} wrapper for executing BOLT-LMM and reading the results.
}
