% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_emmax_ps.R
\name{read_emmax_ps}
\alias{read_emmax_ps}
\title{Read EMMAX statistics table}
\usage{
read_emmax_ps(file, maf = NULL, ext = "ps", verbose = TRUE)
}
\arguments{
\item{file}{The file path to read, possibly missing the extension.}

\item{maf}{Optional vector of allele frequencies (misnamed, these cannot be \emph{minor} allele frequencies or nothing gets corrected), used to correct signs of regression coefficients (EMMAX reports them in terms of the major allele, not the original allele).  If missing, no corrections are applied.}

\item{ext}{The expected file extension.
Set to \code{NA} to prevent adding any extension to \code{file}.
The default "ps" extension matches what EMMAX generates.}

\item{verbose}{If \code{TRUE} (default) function reports the path of the file being loaded (after autocompleting the extensions).}
}
\value{
The table as a \code{tibble}.
The original does not have a header line.
Output has 3 columns named: "id", "beta", "p".
}
\description{
Parses the output generated by EMMAX  (genetic assocation test).
Returned table has standard column names.
}
\examples{
\dontrun{
data <- read_emmax_ps( file, maf )
}

}
\seealso{
\code{\link[=emmax_lmm]{emmax_lmm()}} wrapper for executing EMMAX and reading the results.
}
