% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcta_hsq.R
\name{read_gcta_hsq}
\alias{read_gcta_hsq}
\title{Read GCTA REML variance components table}
\usage{
read_gcta_hsq(file, ext = "hsq", verbose = TRUE)
}
\arguments{
\item{file}{The file path to read, possibly missing the extension.}

\item{ext}{The expected file extension.
Set to \code{NA} to prevent adding any extension to \code{file}.
The default extension "hsq" agrees with the output of \code{\link[=gcta_reml]{gcta_reml()}}.}

\item{verbose}{If \code{TRUE} (default) function reports the path of the file being loaded (after autocompleting the extensions).}
}
\value{
A list containing these named elements:
\itemize{
\item \code{herit}: the estimated heritability.
\item \code{se}: the standard error of the heritability estimate.
\item \code{data}: the full table of variance component estimates and other miscelaneous info.
}
}
\description{
Parses the output generated by GCTA with option \code{--reml} (heritability estimation).
Returned table has column names standardized for convenience.
Validated with GCTA version 1.93.2 beta.
}
\examples{
\dontrun{
obj <- read_gcta_hsq( file )
obj$herit
}

}
\seealso{
\code{\link[=gcta_reml]{gcta_reml()}} wrapper for executing GCTA REML and reading the results.
}
