% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_gcta_mlma.R
\name{read_gcta_mlma}
\alias{read_gcta_mlma}
\title{Read GCTA MLMA statistics table}
\usage{
read_gcta_mlma(file, ext = "mlma", verbose = TRUE)
}
\arguments{
\item{file}{The file path to read, possibly missing the extension.}

\item{ext}{The expected file extension.
Set to \code{NA} to prevent adding any extension to \code{file}.
The default extension "mlma" agrees with the output of \code{\link[=gcta_mlma]{gcta_mlma()}}.}

\item{verbose}{If \code{TRUE} (default) function reports the path of the file being loaded (after autocompleting the extensions).}
}
\value{
The table as a \code{tibble}.
Original names (as they appear in the header line of the file) are modified by lowercasing all, followed by these specific mappings to ensure that columns shared by BIM table have the same names as those returned by \code{\link[genio:read_bim]{genio::read_bim()}}, the association p-value is "p", and the regression coefficient is "beta" ("orig" -> "new"):
\itemize{
\item "snp" -> "id"
\item "bp" -> "pos"
\item "a1" -> "ref"
\item "a2" -> "alt"
\item "b" -> "beta"
}
}
\description{
Parses the output generated by GCTA with option \code{--mlma} (genetic assocation test).
Returned table has column names standardized for convenience.
Validated with GCTA version 1.93.2 beta.
}
\examples{
\dontrun{
data <- read_gcta_mlma( file )
}

}
\seealso{
\code{\link[=gcta_mlma]{gcta_mlma()}} wrapper for executing GCTA MLMA and reading the results.
}
