% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_glm.R
\name{read_plink_glm}
\alias{read_plink_glm}
\title{Read \code{plink2 --glm} association table}
\usage{
read_plink_glm(
  file,
  pheno = "PHENO1",
  ext = "glm.linear",
  verbose = TRUE,
  v = "a6LM"
)
}
\arguments{
\item{file}{The file path to read, missing the extension and phenotype identifier.}

\item{pheno}{An identifier of the phenotype used in file path.
The default "PHENO1" is for the first, unnamed phenotype in a FAM table.}

\item{ext}{The expected file extension.
(In this case, you cannot set to \code{NA} to prevent adding any extension to \code{file}.)
The default extension "glm.linear" agrees with the output of \code{\link[=plink_glm]{plink_glm()}}.}

\item{verbose}{If \code{TRUE} (default) function reports the path of the file being loaded (after autocompleting the extensions).}

\item{v}{Code for format version.
Only special value is 'a3LM', which was a previously validated version (full: "v2.00a3LM AVX2 Intel (4 Aug 2021)") with a certain column specification, but which does not work for latest version.
Any other value sets column specifications to those of "v2.00a6LM AVX2 Intel (4 May 2024)".}
}
\value{
The table as a \code{tibble}.
Original names (as they appear in the header line of the file) are modified by lowercasing all, followed by these specific mappings to ensure that columns shared by BIM table have the same names as those returned by \code{\link[genio:read_bim]{genio::read_bim()}}, the association p-value is "p", and the regression coefficient is "beta" ("orig" -> "new"):
\itemize{
\item "#chrom" -> "chr"
}
}
\description{
Parses the output generated by \code{plink2} with option \code{--glm} (genetic assocation test).
Returned table has column names standardized for convenience.
Validated with plink2 version "v2.00a6LM AVX2 Intel (4 May 2024)".
}
\examples{
\dontrun{
data <- read_plink_glm( file )
}

}
\seealso{
\code{\link[=plink_glm]{plink_glm()}} wrapper for executing \code{plink2 --glm} and reading the results.
}
