% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_plink_hardy.R
\name{read_plink_hardy}
\alias{read_plink_hardy}
\title{Read \code{plink2 --hardy} output table}
\usage{
read_plink_hardy(file, ext = "hardy", verbose = TRUE)
}
\arguments{
\item{file}{The file path to read, missing the extension.}

\item{ext}{The expected file extension.
(In this case, you cannot set to \code{NA} to prevent adding any extension to \code{file}.)
The default extension "hardy" agrees with the output of \code{\link[=plink_hardy]{plink_hardy()}}.}

\item{verbose}{If \code{TRUE} (default) function reports the path of the file being loaded (after autocompleting the extensions).}
}
\value{
The table as a \code{tibble}.
Original names (as they appear in the header line of the file) are modified by lowercasing all, followed by these specific mappings to ensure that columns shared by BIM table have the same names as those returned by \code{\link[genio:read_bim]{genio::read_bim()}}, and the test p-value is "p" ("orig" -> "new"):
\itemize{
\item "#chrom" -> "chr"
\item "a1" -> "ref"
\item "ax" -> "alt"
\item "midp" -> "p" if present (in outputs with option "midp" this column replaces "p")
}
}
\description{
Parses the output generated by \code{plink2} with option \code{--hardy} (Hardy-Weinberg test).
Returned table has column names standardized for convenience.
Validated with plink2 version "v2.00a3LM AVX2 Intel (4 Aug 2021)".
}
\examples{
\dontrun{
data <- read_plink_hardy( file )
}

}
\seealso{
\code{\link[=plink_hardy]{plink_hardy()}} wrapper for executing \code{plink2 --hardy} and reading the results.
}
