% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/require_files_tped.R
\name{require_files_tped}
\alias{require_files_tped}
\title{Require that (old) plink transposed files are present}
\usage{
require_files_tped(file)
}
\arguments{
\item{file}{The shared file path (excluding extensions \code{tped}, \code{tfam}).}
}
\value{
Nothing
}
\description{
This function checks that each of the Plink transposed files (TPED/TFAM extensions) are present, given the shared base file path, stopping with an informative message if any of the files is missing.
This function aids troubleshooting, as various downstream external software report missing files differently and sometimes using confusing or obscure messages.
}
\examples{
\dontrun{
# to require "data.tped" and "data.tfam", run like this:
# (stops if any of the files is missing)
require_files_tped("data")
}

}
\seealso{
\code{\link[=plink1_bed_to_tped]{plink1_bed_to_tped()}} for creating TPED/TFAM files from BED/BIM/FAM files.

\code{\link[=emmax_kin]{emmax_kin()}}, \code{\link[=emmax_lmm]{emmax_lmm()}} for functions that require TPED/TFAM files.
}
